make the args better
